Provide Authority.Ecto.Enum to further reduce the amount of generated code/tests
